+++ react/ProductComparisonWrapper.tsx
@@ -1,4 +1,5 @@
 import React, { ReactChildren, ReactChild } from 'react'
+
 import ComparisonContext from './ProductComparisonContext'
 
 interface Props {
@@ -6,6 +7,7 @@ interface Props {
 }
 const ProductComparisonWrapper = ({ children }: Props) => {
   const { ProductComparisonProvider } = ComparisonContext
+
   return <ProductComparisonProvider>{children}</ProductComparisonProvider>
 }