+++ react/typings/vtex.list-context.d.ts
@@ -1,0 +1,3 @@
+declare module 'vtex.list-context' {
+  export const useListContext: () => undefined | { list: any[] }
+}